xtensa/esp32s3: contain any user fault, not a list of causes
The user-exception handler recovered from three causes -- Load, Store and
InstrFetch Prohibited -- plus an illegal instruction, and panicked on
everything else.  Gating on a list of causes means every cause left off the
list is a way for an unprivileged task to stop the machine, and a task has
plenty:  a divide by zero (EXCCAUSE 6), a privileged instruction (8), a
load/store error (3), an instruction-fetch error (2), a data access to an
address the bus will not carry (9).  None of these involve a boundary at all;
they are things a correctly confined application does to itself.

Measured on an ESP32-S3-WROOM-2, kernel build, before this change:  a user
task dividing by zero takes the whole system down, silently.

So gate on the interruptee instead.  The dispatcher still gets first refusal
on 28/29/20, which are the precise, restartable causes and the only ones
re-executing can help.  Anything it does not service now goes to the abort
path if the saved PS says the fault was taken in User Mode.  That bit is also
what excludes the cases with no safe task to kill:  a kernel thread, a fault
inside a system call made on the user's behalf, and a fault while handling an
interrupt all run with PS.UM clear and still panic.  This is the same shape
riscv_fault_handler() has, where the cause is used for the message only.

The abort itself moves to esp32s3_userfault.c and stops depending on the
recoverable-fault dispatcher.  CONFIG_ESP32S3_PAGEFAULT_ABORT was a child of
CONFIG_ESP32S3_PAGEFAULT, which is default n, so "do not let a rogue task
halt the machine" was unreachable unless you also opted into "retry the
faulting instruction".  Those are separate capabilities and the safety one is
the one you always want, so it becomes CONFIG_ESP32S3_USERFAULT_ABORT,
default y wherever there is an unprivileged world.  A BUILD_PROTECTED
configuration previously got neither.

Verified on hardware with the new examples/misbehave.  A user task that
writes through NULL (29), reads a wild address (28), divides by zero (6) or
calls into a buffer of garbage (20) is terminated on its own, with an
unrelated task still counting either side of it.

Two limits worth recording rather than hiding:

* Stack overflow is not contained and does not even report -- the console
  goes silent.  On a windowed ABI the overflow faults inside the window
  overflow handler, so it arrives as a double exception with PS.UM already
  clear and no gate can see it.  Guard pages are the answer and are what
  CONFIG_ESP32S3_PAGEFAULT exists to make possible; that is separate work.
* A 32-bit load from an unaligned address does not trap on this core, so
  EXCCAUSE 9 is reachable by other means but not that one.


(cherry picked from commit 3c11f3793dbae918dc6ebe3e99cbb60f70648b3e)
