+++ arch/xtensa/src/esp32s3/esp32s3_pagefault.c
@@ -34,15 +34,8 @@
 #include <arch/irq.h>
 #include <arch/xtensa/xtensa_corebits.h>
 
-#ifdef CONFIG_ESP32S3_PAGEFAULT_ABORT
-#include <signal.h>
-#endif
-
 #include "xtensa.h"
 #include "sched/sched.h"
-#ifdef CONFIG_ESP32S3_PAGEFAULT_ABORT
-#include "signal/signal.h"
-#endif
 
 /****************************************************************************
  * Pre-processor Definitions
@@ -184,60 +177,19 @@ int esp32s3_pagefault_dispatch(int exccause, uint32_t *regs)
   return -EFAULT;
 }
 
-#ifdef CONFIG_ESP32S3_PAGEFAULT_ABORT
 /****************************************************************************
- * Name: esp32s3_pagefault_abort
+ * Name: esp32s3_pagefault_clear_repeat
  *
  * Description:
- *   Terminate just the faulting unprivileged (WORLD1) task by delivering a
- *   fatal SIGSEGV, instead of panicking the whole system.  Called from
- *   xtensa_user() for an unrecoverable user-mode cache-attribute fault.
- *
- *   Mirrors the interrupt-dispatch handshake: record the exception frame as
- *   the task context, dispatch the signal -- which redirects the task to the
- *   signal trampoline via up_schedule_sigaction() -- then return the
- *   redirected frame so the vector's RFE resumes the task in the trampoline,
- *   whose SIGSEGV default action (_exit) tears the task down and resumes.
- *
- * Returned Value:
- *   The register frame to resume (the signal trampoline for the faulting
- *   task).
+ *   Forget the last serviced fault.  Called once a fault has been contained
+ *   some other way -- the task terminated -- so that later, unrelated faults
+ *   at the same address are not counted as runaway recursion.
  *
  ****************************************************************************/
 
-uint32_t *esp32s3_pagefault_abort(int exccause, uint32_t *regs)
+void esp32s3_pagefault_clear_repeat(void)
 {
-  struct tcb_s *tcb = this_task();
-  siginfo_t     info;
-
-  /* Reaching here means the fault was contained and the system carried on,
-   * so the repeat counter has served its purpose.  Clear it, or a probe run
-   * three times at one address would trip the guard in the dispatcher and
-   * halt a perfectly healthy system.  Only *unbroken* recursion -- a report
-   * that faults before the abort can happen -- should stop the machine.
-   */
-
   g_pf_last_vaddr = 0;
   g_pf_last_pc    = 0;
   g_pf_repeats    = 0;
-
-  _alert("SIGSEGV task %s: EXCCAUSE=%d EXCVADDR=%08x PC=%08x\n",
-         get_task_name(tcb), exccause, (unsigned)regs[REG_EXCVADDR],
-         (unsigned)regs[REG_PC]);
-
-  up_set_interrupt_context(true);
-  tcb->xcp.regs = regs;
-
-  info.si_signo           = SIGSEGV;
-  info.si_code            = SI_USER;
-  info.si_errno           = 0;
-  info.si_value.sival_ptr = (FAR void *)regs[REG_EXCVADDR];
-
-  nxsig_tcbdispatch(tcb, &info, false);
-
-  regs          = tcb->xcp.regs;
-  tcb->xcp.regs = NULL;
-  up_set_interrupt_context(false);
-  return regs;
 }
-#endif /* CONFIG_ESP32S3_PAGEFAULT_ABORT */
