+++ arch/xtensa/src/esp32s3/esp32s3_user.c
@@ -31,6 +31,7 @@
 #ifdef CONFIG_ESP32S3_PAGEFAULT
 #include "esp32s3_pagefault.h"
 #endif
+#include "esp32s3_userfault.h"
 #ifdef CONFIG_ESPRESSIF_SPIFLASH
 #include "esp_private/cache_utils.h"
 #endif
@@ -98,41 +99,39 @@ uint32_t *xtensa_user(int exccause, uint32_t *regs)
         {
           return regs;
         }
-
-#ifdef CONFIG_ESP32S3_PAGEFAULT_ABORT
-      /* Unrecoverable, but a WORLD1 (user-mode) fault need not take down the
-       * whole system: terminate just the faulting task with SIGSEGV.  The
-       * User Mode bit in the interruptee's saved PS distinguishes a user
-       * fault from a privileged one (which still panics below).
-       */
-
-      if ((regs[REG_PS] & PS_UM) != 0)
-        {
-          return esp32s3_pagefault_abort(exccause, regs);
-        }
-#endif
     }
 #endif
 
-#if defined(CONFIG_ESP32S3_PAGEFAULT) && defined(CONFIG_ESP32S3_PAGEFAULT_ABORT)
-  /* A *denied* external-memory access does not trap on this chip.  TRM v1.8
-   * p.699: an access without permission is "responded with 0 (for internal
-   * memory) or 0xdeadbeaf (for external memory)".  So an unprivileged branch
-   * into kernel text in flash or PSRAM is refused silently, the CPU executes
-   * the dummy word it was handed instead, and the refusal surfaces here as
-   * EXCCAUSE_ILLEGAL at the address that was branched to -- never as
-   * EXCCAUSE_INSTR_PROHIBITED.
+#ifdef CONFIG_ESP32S3_USERFAULT_ABORT
+  /* Nothing serviced it, so this fault is not going away by re-executing.
+   * If the interruptee was unprivileged, terminate just that task.
    *
-   * Re-executing cannot help: the instruction genuinely is not there, so
-   * this is not offered to the dispatcher above.  Terminate just the task.
-   * That is the right answer whichever way the illegal instruction arose --
-   * a refused fetch or simply a corrupt user binary -- because a user task
-   * running garbage must not take the system down with it.
+   * The test is on the interruptee's User Mode bit and NOT on the cause, and
+   * that is the whole point.  A user task has many ways to raise a
+   * synchronous exception that no dispatcher can service -- an unaligned
+   * load (EXCCAUSE 9), a divide by zero (6), a privileged instruction (8), a
+   * load/store error (3), a corrupt opcode (0) -- and gating on a list of
+   * causes means every cause left off the list is a way for an unprivileged
+   * task to stop the machine.  Whichever way it arose, a user task running
+   * garbage must not take the system down with it.
+   *
+   * PS.UM also excludes the cases where there is no safe task to kill:  a
+   * kernel thread, a fault inside a system call made on the user's behalf,
+   * and a fault while handling an interrupt all run with it clear, and fall
+   * through to the panic below.
+   *
+   * One cause deserves its own note.  A *denied* external-memory access does
+   * not trap on this chip.  TRM v1.8 p.699: an access without permission is
+   * "responded with 0 (for internal memory) or 0xdeadbeaf (for external
+   * memory)".  So an unprivileged branch into kernel text in flash or PSRAM
+   * is refused silently, the CPU executes the dummy word it was handed
+   * instead, and the refusal surfaces here as EXCCAUSE_ILLEGAL at the
+   * address that was branched to -- never as EXCCAUSE_INSTR_PROHIBITED.
    */
 
-  if (exccause == EXCCAUSE_ILLEGAL && (regs[REG_PS] & PS_UM) != 0)
+  if ((regs[REG_PS] & PS_UM) != 0)
     {
-      return esp32s3_pagefault_abort(exccause, regs);
+      return esp32s3_userfault_abort(exccause, regs);
     }
 #endif
